Allow examsys-docker.sh to be run from anywhere
Before if you tried to run examsys-docker.sh when you were not in the same directory as the script it would fail, because docker compose would look for the yml files were in your working directory.

This change includes them using their full path, so they will not be looked for in your working directory.